backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
requests
psycopg2-binary
PyJWT
Flask-CORS
redis
rq
